regulation type to float

+++ ADAUGDB/views/course_structure.py
@@ -29,7 +29,7 @@ def add_course_structre(request):
                 errorData = []
                 newDataset.headers = ['BYear', 'BSem', 'Dept', 'Regulation', 'Category', 'Type', 'Creditable', 'Credits', 'count']
                 for row in dataset:
-                    if row[3] == int(form.cleaned_data.get('regulation')):
+                    if row[3] == float(form.cleaned_data.get('regulation')):
                         newRow = (row[0], row[1], row[2], row[3], row[4], row[5], row[6], row[7], row[8])
                         newDataset.append(newRow)
                     else:
